Refactoring deserializer and tests.
