+++ tests/myapp/tests/__init__.py
@@ -1,0 +1,2 @@
+from .serializer import *
+from .deserializer import *
